AIModel/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
torch
torchvision
pillow
numpy
huggingface_hub>=0.21.0
python-multipart>=0.0.9
